Rmarkdown Template for koma-scrartcl

This template is a designed for pdf output and acts as an alternative to the tufte::tufte_handout() template. Whereas that Rmarkdown template is based on the tufte-latex LaTeX package, owever, komatufte is instead based on a custom template that I have designed using the KOMA-script scrartcl class. Most of the added functionality is documented in the komatufte vignette (vignette("komatufte", "komadown")) and is accessed via the YAML metadata block of the R Markdown file.

komatufte(..., number_sections = TRUE,
  md_extensions = c("-autolink_bare_uris"), csl = NULL,
  colorlinks = TRUE)

Arguments

...
arguments passed down to bookdown::pdf_document2()

number_sections
Whether to number section headers: if TRUE, figure/table numbers will be of the form X.i, where X is the current first-level section number, and i is an incremental number (the i-th figure/table); if FALSE, figures/tables will be numbered sequentially in the document from 1, 2, ..., and you cannot cross-reference section headers in this case.

md_extensions
Arguments to be passed to a specific output format function. For a function foo2(), its arguments are passed to foo(), e.g. ... of html_document2() are passed to rmarkdown::html_document().

csl
bibliography style in the .csl format

colorlinks
should cross-references and links be colored?

Value

R Markdown output format to pass to rmarkdown::render()

Details

The main features of this template is the use of a left-aligned column for the main content and a right-aligned column for marginal content.

This template is more opinionated than scrartcl and is not catered to manual tinkering with the style.

Examples

library(rmarkdown)
library(bookdown)
rmarkdown::draft("my-article.Rmd",
                 template = "komatufte",
                 package = "komadown",
                 edit = FALSE)

#> Error in rmarkdown::draft("my-article.Rmd", template = "komatufte", package = "komadown",     edit = FALSE): The template 'komatufte' was not found in the komadown package