Rmarkdown Template for koma-scrartcl

This is an interface to the KOMA-Script class scrartcl that acts as an alternative to the article class in LaTeX. Most of the added functionality is documented in the scrartcl vignette (vignette("scrartcl", "komadown")) and is accessed via the YAML metadata block of the R Markdown file.

scrartcl(..., number_sections = TRUE,
  md_extensions = c("-autolink_bare_uris"),
  csl = find_resource("scrartcl", "vancouver-brackets.csl"))

Arguments

...
Arguments passed down to bookdown::pdf_document2()

number_sections
Whether to number section headers: if TRUE, figure/table numbers will be of the form X.i, where X is the current first-level section number, and i is an incremental number (the i-th figure/table); if FALSE, figures/tables will be numbered sequentially in the document from 1, 2, ..., and you cannot cross-reference section headers in this case.

md_extensions
Arguments to be passed to a specific output format function. For a function foo2(), its arguments are passed to foo(), e.g. ... of html_document2() are passed to rmarkdown::html_document().

csl
bibliography style in the .csl format

Value

R Markdown output format to pass to rmarkdown::render().

Examples

library(rmarkdown)
library(bookdown)
draft("MyArticle.Rmd", template = "scrartcl", package = "komadown",
      edit = FALSE)

#> Error in draft("MyArticle.Rmd", template = "scrartcl", package = "komadown",     edit = FALSE): The template 'scrartcl' was not found in the komadown package